requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
confusable-homoglyphs==3.3.1
cryptography==43.0.0
defusedxml==0.7.1
dj-database-url
Django==5.0.7
django-allauth==0.63.6
django-photologue==3.17
django-registration==3.4
django-sortedm2m==3.1.1
django-userprofile==0.0.6
ExifRead==3.0.0
idna==3.7
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-http-client==3.3.7
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sendgrid==3.6.5
sendgrid-django==4.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
